Fix tests WIP
